+++ README.md
@@ -1,3 +1,12 @@
-# BayesOptGH
+# Tunny
 
-WIP
+**Tunny** はオープンソースのハイパーパラメータ自動最適化フレームワーク Optuna を使った Grasshopper の最適化コンポーネントです。
+以下は公式サイトより引用
+
+> オープンソースのハイパーパラメータ自動最適化フレームワーク Optuna™ は、ハイパーパラメータの値に関する試行錯誤を自動化し、優れた性能を発揮するハイパーパラメータの値を自動的に発見します。オープンソースの深層学習フレームワーク Chainer をはじめ、様々な機械学習ソフトウェアと一緒に使用することが可能です。
+> 
+> Optuna は、物体検出コンペティション [Google AI Open Images 2018 – Object Detection Track](https://www.preferred.jp/ja/news/pr20180907/) など、PFN の各プロジェクトで活用され、成果をあげています。
+
+
+Optuna の公式サイト
+- https://optuna.org/
