+++ Tunny.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31729.503
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BayesOpt", "BayesOpt\BayesOpt.csproj", "{623ABE73-A37F-4E26-AD90-B5C55AB4AFFA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tunny", "Tunny\Tunny.csproj", "{623ABE73-A37F-4E26-AD90-B5C55AB4AFFA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
