+++ Tunny/UI/OptimizationWindow.resx
